test: migrate ColorValue, NumberValue, StringValue, UnitValue
